test: extra test case
