+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/utils/utils.kt
@@ -64,7 +64,7 @@ internal fun <E> countsToList(counts: Map<E, Int>): List<E> {
  * @param counts [CountsMap]<E>: counts map
  * @param multiset [MultiSet]<E>: MultiSet to combine with
  * @param operation (Int, Int) -> Int: combination function
- * @param useAllValues [Boolean]: if all values from both sets should be used to generate the new set. If `false`, only the values from the counts map will be used.
+ * @param useAllValues [Boolean]: if all values from the map and the set should be used to generate the new set. If `false`, only the values from the counts map will be used.
  * @param const [Boolean]: if the returned MultiSet should be a ConstMultiSet
  * @return [MultiSet]<E>: new set where each element has the number of occurrences specified by the operation. If [const] is `true`, the set will be a const multi set.
  * Defaults to `false`.
